Move getConfig into test/
This has only been in use by tests since e37e1a6c203f1e85005831a2bc874aa0b800e3ad, so doesn't belong in the src/ folder anymore
